images

+++ app/Http/Controllers/Api/Auth/UpdateUserProfileController.php
@@ -12,6 +12,7 @@ use App\Http\Controllers\Controller;
 use App\Rules\BooleanAttribute;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
+use Intervention\Image\ImageManagerStatic as Image;
 
 /**
  * @group User Profile
@@ -42,7 +43,7 @@ class UpdateUserProfileController extends Controller
             'receive_emails' => ['required', 'boolean'],
             'receive_push_notifications' => ['required', 'boolean'],
             'default_distance_unit' => ['required', 'string', 'in:kilo,mile'],
-            'image' => ['nullable', 'mimes:jpeg,jpg,png,gif', 'max:5102'],
+            'image' => ['sometimes', 'base64dimensions:min_width=100,min_height=200'],
         ]);
 
         if ($validate_request->fails()) {
@@ -61,10 +62,16 @@ class UpdateUserProfileController extends Controller
         ]);
 
         if ($request->has('image') && $request->image !== '' && !is_null($request->image)) {
-            Storage::disk('public')->delete($user->image);
+
+            if ($user->image != 'images/profile/default-profile.png') {
+                Storage::disk('public')->delete($user->image);
+            }
+            $image_name = \Str::random(15) . '.' . 'png';
+            $path = public_path('/images/profile/' . $image_name);
+            Image::make(file_get_contents($request->image))->save($path);
 
             $user->update([
-                'image' => $request->file('image')->store('images/profile')
+                'image' =>   'images/profile/' . $image_name
             ]);
         }
